feat: implement mobile dashboard menu and update navbar routing for provider navigation

+++ src/graphql/mutations/reviews.js
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const ADD_REVIEW_MUTATION = gql`
-  mutation AddReview($bookingId: ID!, $rating: Float!, $comment: String) {
+  mutation AddReview($bookingId: ID!, $rating: Int!, $comment: String) {
     addReview(bookingId: $bookingId, rating: $rating, comment: $comment) {
       id
       rating
